+++ src/main/java/com/telran/shopmongodb/controller/dto/AddUserBalanceResponseDto.java
@@ -12,5 +12,5 @@ import java.math.BigDecimal;
 @Builder
 public class AddUserBalanceResponseDto {
     String email;
-    BigDecimal balance;
+    Double balance;
 }
